# app_backend/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "ConversationPage": {
        "additionalProperties": false,
        "properties": {
          "app_user_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "App User Id",
            "type": "integer"
          },
          "day": {
            "title": "Day",
            "type": "string"
          },
          "dog_user_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Dog User Id",
            "type": "integer"
          },
          "id": {
            "anyOf": [
              {
                "maximum": 9007199254740991.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageEntry"
            },
            "title": "Messages",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "id",
          "app_user_id",
          "dog_user_id",
          "day",
          "messages"
        ],
        "title": "ConversationPage",
        "type": "object"
      },
      "DispatchReceipt": {
        "additionalProperties": false,
        "properties": {
          "conversation_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Conversation Id",
            "type": "integer"
          },
          "day": {
            "title": "Day",
            "type": "string"
          },
          "message_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Message Id",
            "type": "integer"
          },
          "request_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Request Id",
            "type": "integer"
          },
          "status": {
            "const": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "conversation_id",
          "message_id",
          "request_id",
          "day",
          "status"
        ],
        "title": "DispatchReceipt",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageEntry": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Id",
            "type": "integer"
          },
          "reply_to": {
            "anyOf": [
              {
                "maximum": 9007199254740991.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply To"
          },
          "request_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Request Id",
            "type": "integer"
          },
          "role": {
            "enum": [
              "app_user",
              "robot",
              "resident"
            ],
            "title": "Role",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "queued",
                  "accepted",
                  "completed",
                  "failed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "text",
          "timestamp",
          "request_id"
        ],
        "title": "MessageEntry",
        "type": "object"
      },
      "MessageReply": {
        "additionalProperties": false,
        "properties": {
          "final": {
            "default": true,
            "title": "Final",
            "type": "boolean"
          },
          "reply_to": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Reply To",
            "type": "integer"
          },
          "request_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Request Id",
            "type": "integer"
          },
          "role": {
            "enum": [
              "robot",
              "resident"
            ],
            "title": "Role",
            "type": "string"
          },
          "text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "reply_to",
          "role",
          "text",
          "timestamp"
        ],
        "title": "MessageReply",
        "type": "object"
      },
...
      "NewMessage": {
        "additionalProperties": false,
        "properties": {
          "app_user_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "App User Id",
            "type": "integer"
          },
          "text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "app_user_id",
          "text"
        ],
        "title": "NewMessage",
        "type": "object"
      },
      "NewNote": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "occurrence_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Occurrence Id",
            "type": "integer"
          },
          "outcome": {
            "enum": [
              "completed",
              "not_completed",
              "unknown"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "reminder_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Reminder Id",
            "type": "integer"
          },
          "request_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Request Id",
            "type": "integer"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "occurrence_id",
          "reminder_id",
          "timestamp",
          "description",
          "outcome"
        ],
        "title": "NewNote",
        "type": "object"
      },
      "NewNotification": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "dog_user_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Dog User Id",
            "type": "integer"
          },
          "is_emergency": {
            "title": "Is Emergency",
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "dog_user_id",
          "timestamp",
          "description",
          "is_emergency"
        ],
        "title": "NewNotification",
        "type": "object"
      },
...
      "Note": {
        "additionalProperties": false,
        "properties": {
          "day": {
            "title": "Day",
            "type": "string"
          },
          "description": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "dog_user_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Dog User Id",
            "type": "integer"
          },
          "id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Id",
            "type": "integer"
          },
          "occurrence_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Occurrence Id",
            "type": "integer"
          },
          "outcome": {
            "enum": [
              "completed",
              "not_completed",
              "unknown"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "reminder_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Reminder Id",
            "type": "integer"
          },
          "request_id": {
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Request Id",
            "type": "integer"
          },
          "source": {
            "enum": [
              "robot",
              "seed"
            ],
            "title": "Source",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "timestamp",
          "request_id",
          "occurrence_id",
          "reminder_id",
          "description",
          "outcome",
          "dog_user_id",
          "source",
          "day"
        ],
        "title": "Note",
        "type": "object"
      },
...
      "Page_HistoryItem_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[HistoryItem]",
        "type": "object"
      },
      "Page_Note_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Note"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[Note]",
        "type": "object"
      },
      "Page_Notification_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Notification"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[Notification]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Annie Family API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/history": {
      "get": {
        "operationId": "history_api_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "app_user_id",
            "required": true,
            "schema": {
              "maximum": 9007199254740991,
              "minimum": 1,
              "title": "App User Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_HistoryItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History",
        "tags": [
          "history"
        ]
      }
    },
    "/api/messages": {
      "get": {
        "operationId": "messages_api_messages_get",
        "parameters": [
          {
            "in": "query",
            "name": "app_user_id",
            "required": true,
            "schema": {
              "maximum": 9007199254740991,
              "minimum": 1,
              "title": "App User Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Messages",
        "tags": [
          "messages"
        ]
      },
      "post": {
        "operationId": "new_message_api_messages_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_.:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "New Message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/messages/replies": {
      "post": {
        "operationId": "reply_api_messages_replies_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_.:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageReply"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reply",
        "tags": [
          "robot callbacks"
        ]
      }
    },
    "/api/notes": {
      "get": {
        "operationId": "notes_api_notes_get",
        "parameters": [
          {
            "in": "query",
            "name": "app_user_id",
            "required": true,
            "schema": {
              "maximum": 9007199254740991,
              "minimum": 1,
              "title": "App User Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "reminder_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reminder Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Note_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Notes",
        "tags": [
          "reminders"
        ]
      },
      "post": {
        "operationId": "note_api_notes_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_.:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewNote"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Note",
        "tags": [
          "robot callbacks"
        ]
      }
    },
    "/api/notifications": {
      "get": {
        "operationId": "notifications_api_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "app_user_id",
            "required": true,
            "schema": {
              "maximum": 9007199254740991,
              "minimum": 1,
              "title": "App User Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "is_emergency",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Emergency"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Notification_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Notifications",
        "tags": [
          "history"
        ]
      },
      "post": {
        "operationId": "notification_api_notifications_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_.:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewNotification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Notification",
        "tags": [
          "robot callbacks"
        ]
      }
    },
    "/api/reminders": {
      "get": {
        "operationId": "reminders_api_reminders_get",
        "parameters": [
          {
            "in": "query",
            "name": "app_user_id",
            "required": true,
            "schema": {
              "maximum": 9007199254740991,
              "minimum": 1,
              "title": "App User Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reminders",
        "tags": [
          "reminders"
        ]
      },
      "post": {
        "operationId": "new_reminder_api_reminders_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_.:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewReminder"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reminder"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "New Reminder",
        "tags": [
          "reminders"
        ]
      }
    },
    "/api/requests/{request_id}": {
      "get": {
        "operationId": "request_status_api_requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "app_user_id",
            "required": true,
            "schema": {
              "maximum": 9007199254740991,
              "minimum": 1,
              "title": "App User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Status",
        "tags": [
          "messages"
        ]
      }
    },
...
  }
}